More aurora::gx refactoring

+++ Runtime/Streams/ContainerReaders.hpp
@@ -6,7 +6,7 @@
 namespace metaforce {
 template <class T, size_t N>
 void read_reserved_vector(rstl::reserved_vector<T, N>& v, CInputStream& in) {
-  u32 count = in.ReadLong();
+  u32 count = in.ReadUint32();
   v.resize(count);
   for (u32 i = 0; i < count; ++i) {
     v[i] = in.Get<T>();
@@ -15,7 +15,7 @@ void read_reserved_vector(rstl::reserved_vector<T, N>& v, CInputStream& in) {
 
 template <class T>
 void read_vector(std::vector<T>& v, CInputStream& in) {
-  u32 count = in.ReadLong();
+  u32 count = in.ReadUint32();
   v.reserve(count);
   for (u32 i = 0; i < count; ++i) {
     v.emplace_back(in.Get<T>());
